@@ -3,7 +3,7 @@
 #
 # This class defines the create dummy followees migration in follow system
 ###
-class CreateDummyFollowees < ActiveRecord::Migration
+class CreateDummyFollowees < ActiveRecord::Migration[5.0]
   ###
   # Changes the database
   ###
